Abstract out DailyTable
